fix: type error

+++ alignment/utils.py
@@ -51,7 +51,7 @@ def apply_attention_make_batch(tensor,mask,index,tbtt_step,x_length,y_length):
     return tensor_pad,mask_pad.bool()
 
 def y_make_batch(input_y_tensor,y_tensor,chunk_index,tbtt_step):
-    y_input_pad = torch.IntTensor(input_y_tensor.size(0),tbtt_step)
+    y_input_pad = torch.LongTensor(input_y_tensor.size(0),tbtt_step)
     y_pad = torch.IntTensor(input_y_tensor.size(0),tbtt_step)
 
     for i in range(y_pad.size(0)):
